Further changes and fixing compilation errors

+++ VulkanGeometry.hpp
@@ -22,14 +22,14 @@ namespace VkApplication {
                 Vertex vertex{};
 
                 vertex.pos = {
-                    attrib.vertices[3 * index.vertex_index + 0],
-                    attrib.vertices[3 * index.vertex_index + 1],
-                    attrib.vertices[3 * index.vertex_index + 2]
+                    attrib.vertices[3 * (int)index.vertex_index + 0],
+                    attrib.vertices[3 * (int)index.vertex_index + 1],
+                    attrib.vertices[3 * (int)index.vertex_index + 2]
                 };
 
                 vertex.texCoord = {
-                    attrib.texcoords[2 * index.texcoord_index + 0],
-                    1.0f - attrib.texcoords[2 * index.texcoord_index + 1]
+                    attrib.texcoords[2 * (int)index.texcoord_index + 0],
+                    1.0f - attrib.texcoords[2 * (int)index.texcoord_index + 1]
                 };
 
                 vertex.color = { 1.0f, 1.0f, 1.0f };
